+++ nakadi-types/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.26.0]
+
+### CHANGED
+
+- Result for publishing is now an enum `SubmissionFailure`
+
+### REMOVED
+
+- `BatchResponse` in favour of `FailedSubmission` containing `SubmissionFailure`
+
 ## [0.25.4] - 2020-05-14
 
 ## [0.25.3] - 2020-05-14